More robust configuration search and replace
Why?

The dockerhub images are created by extracting files from a
locally running docker instance. On the local instance,
config-local.sh has already executed and replaced
_kamailio.2600hz.com_ with _kamailio.kazoo_.

When config-local.sh then runs on the image uploaded to dockerhub,
no replacement takes place as it is looking for _kamailio.2600hz.com_
and the file contains _kamailio.kazoo_.

**Note**: This changes the configuration to rely on `$(hostname)` instead
of `$HOSTNAME`, which seems more consistent with how the other images are
configured.